renamed files to better reflect that app, passport are created by creator functions
